fabtests/psm3: Disable fi_rdm_tagged_peek for cleanup failure
fi_rdm_tagged_peek fails to cleanup with "munmap_chunk(): invalid pointer"
when trying to free hfi_nids in psm_ep.c:1161.
This test is successful when FI_PROVIDER is unset and fails when it is
set to "psm3" or "PSM3". There is an open issue in ofiwg/libfabric to track
this bug. When it is resolved we can re-enable this test.

Issue opened: 10123

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>