prov/shm: add return and more descriptive error message
Fix missing return from unexpected SAR path when unexpected buffer pool
ran out of memory. Add more helpful message to warning log for what
happened.
FI_ENOMEM and FI_ETRUNC are not fully handled in the SAR path and
will cause hangs if it occurs, but this is closer to correct

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>