prov/shm: Touch all buffer pages after DSA page fault
Previously, it was up to user to set the optional FI_SHM_ENABLE_DSA_PAGE_TOUCH
flag to enable optimization that reduces DSA page faults. If not set,
performance may take a hit for workloads with high page fault counts. We now
always enable this optimization, and thus remove the flag. Workloads with no
page faults are not affected.

DSA page faults will now always trigger touching all of the pages in the
descriptor buffer (as it is done in dsa_touch_buffer_pages()), instead of just
one individual faulting page at a time (as was done in
dsa_handle_page_fault()).

Signed-off-by: Jorge Cabrera <jorge.cabrera@intel.com>