fabtests: Address integer overflow coverity issues in fabtests
There are 4 integer overflow issues in fabtests. All of them
need a check that ensures integer overflow does not oocur.

1. changes to harness.c:
This modification includes a check to ensure that adding ret
to m does not exceed SIZE_MAX.

2. changes to shared.c:
This modification includes a check before the addition to
ensure that (rcvd + ret) or (sent + ret) does not exceed SIZE_MAX.
If the addition would exceed SIZE_MAX, it sets err to an error
value and breaks from the loop, preventing overflow.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>