libfabric/core: Fix overflow issue in ofi_rbinit
This modification adds a check for a zero or excessively
large size input, which could lead to rb->size being zero
after the call to roundup_power_of_two. It also ensures
that roundup_power_of_two does not return 0 unless the input
is 0, which should be handled by the initial check.
This approach prevents the underflow that would cause an
integer overflow when rb->size - 1 is calculated.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>