prov/verbs: Replace __BITS_PER_LONG with LONG_WIDTH
The verbs provider uses asm/types.h for __BITS_PER_LONG, but that header
is limited to Linux systems. LONG_WIDTH is defined in C23 in
limits.h. Use it instead, and add logic to define it if not
available. This will allow the verbs provider to compile on FreeBSD
systems.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>