include/rdma, prov/shm,efa,util: update and clarify peer SRX API flow
This is a set of changes aimed at simplifying and clarifying
the peer SRX flow:

- Change fi_peer_rx_entry->size field to "msg_size" to clarify
distinction between size of posted receive and size of incoming
message. Wording is added to the man page to clarify the use of
this field and that the peer can expect it to be initialized
after returning from the owner_op matching calls
- Change get_msg and get_tag to take in a single attributes paramter
instead of separate fields for addr, size, and tag. This simplifies
the matching call and allows for future extension if needed
- Update shm and efa providers with these updates as well as the util_srx
implementation

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>