prov/util: Allow providers to update cache MR IOV
A provider may update the memory region that is added to accommodate
for instance alignment of the region to a larger page boundary. In such
cases, the MR cache info used to search the cache should use the updated
region.

This allows the provider to avoid walking /proc/pid/smaps if the
underlying kernel component may more efficiently determine the backing
page size.

Signed-off-by: Steve Welch <welch@hpe.com>
Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>