include/rdma: add FI_PEER as a capability
The FI_PEER bit is already defined as a flag for the calls to import
shared resources into peers. Extend this flag to be defined as a
primary capability which indicates that the provider supports being
used as a peer through the peer API flow. The subset of the peer API
defined for the FI_PEER capability is the minimum needed for peer support:
CQ, counter, and shared receive queues

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>