core: Introduce Sub-MR
Memory registration consists of two parts: map/pin the memory for local
access and export with a key for remote access. The first part is usually
heavyweight and requries kernel involvement. The second part is less
expensive and can be further separated into key allocation and key
assignment. Key allocation may needs kernel involvement, but key assignment
can be done in user space. Here sub-MR is introduced as a way to allow
separattion of the forementioned two parts, and key reservation is added
to further optimize sub-MR creation.

A sub-MR is created from an existing MR (the base MR).  It inherits the
memory mapping/pinning of the base MR but has its own access key.  The
address range exposed can be same as the the base MR or a subpart of that.
The access rights can be different, too.

Now the base MR can be created with a few extra keys reserved. These reserved
keys will be automatically used for sub-MR registration. This only applies
to FI_MR_PROV_KEY mode.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>