prov/shm: refactor initialization process
The id field was incorrectly being used as an indicator that
a region was mapped. Instead a valid id shows that a peer is
in the rbmap and not that region memory is valid.

Map_to_region has moved out of map_add and a region can only be
initialized in verify_peer or in progress_connreq. This was done
to reduce initialization time and memory footprint because there
are many cases where communication doesn't happen across every
peer. We only need to initialize a region when we need to
communicate with that peer.

Previously map_to_endpoint would be able to map to an endpoint
if the peer was in the map but the region wasn't mapped. This
now protects against segmentation faults because the region must
be valid. Since map_to_region was moved out of map_add there is
no point in doing map_to_endpoint on av_insert because the region
is not mapped. Map_to_endpoint now happens at the end of
map_to_region because we guarentee that the peer is in the map
and its region is valid. This endpoint mapping does not need to
happen for every possible peer so we can use a for each call to
do it for initialized peers only.

With all of these changes we can silence coverity issues about
accessing the fields in the map without a lock. This is because
these fields are properly protected and region and id are
guaranteed to be valid at the point of accessing them.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>