prov/shm: Put smr_map memory into av
Move smr_map_create to an initialization type function.
The map doesn't need to be calloc'd because there is a
1:1 relationship between av:map so there is no reason to
allocate them separately. With this change these functions
can move into smr_av.c because that is the only place they
are called.

peer_data.addr.name is not used anywhere and is already
initialized to zeros on region creation so it doesn't
need to be initialized.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>