fabtests/efa: Avoid testing duplicate mixed memory type workload
Currently, we always test mixed memory type workload in two directions explicitly:
like cuda-to-host, host-to-cuda.

However. pingpong and rma read/write test are all bi-directional, which means there is
no difference for cuda-to-host and host-to-cuda. Running them twice in this situation is
not necessary and causing longer test duration.

This patch improves this by having a reduced memory type list for bi-directional tests.

Signed-off-by: Shi Jin <sjina@amazon.com>