prov/efa: Shrink the size of extra_info array
Currently, extra_info is a 256*8 bytes array in
efa_rdm_ep and efa_rdm_peer, this is way too much
as it can cover 256*64 bits of extra features/requests
that RDM endpoint is likely to have.
This patch reduces it to 4*8 bytes array that
can cover 256 bits which should be enough for the foreseeable
future.

Signed-off-by: Shi Jin <sjina@amazon.com>