core: Remove CURRENT_SYMVER() macro
CURRENT_SYMVER() is just a shortcut for DEFAULT_SYMVER() with version
set to CURRENT_ABI. Using DEFAULT_SYMVER() directly with explict
version string makes the version info more visible from the code and
eliminates the need to change from CURRENT_SYMVER() to DEFAULT_SYMVER()
when the ABI version is updated in the future.

Remove the definition of CURRENT_SYMVER() and CURRENT_ABI from the
header. Update the documentation accordingly.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>