prov/shm: use owner-allocated srx
The peer API has been updated to specify that the owner must allocate
the peer's fid_peer_srx. The shm implementation was allocating its
own internal fid_peer_srx.
This updates the shm implementation to assume it has a unique
fid_peer_srx and updates the imported fid_peer_srx peer_ops, saving
a pointer to the fid_peer_srx instead of the internal fid_ep which
required a wrapper function to get back to the fid_peer_srx

It also returns an internal fid_ep for the created srx which is used
to close the srx by the owner. Even though shm doesn't need anything
attached to the internal fid_ep, it is there for consistency and to
track the domain reference counting for errors.

This patch also moves the srx specific functions into smr_domain
where they belong

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>