prov/efa/test: Use correct qp num in the mock
Currently, the test uses hard-coded 0 as the return
value of efa_mock_ibv_read_qp_num_return_mock, which is
wrong as the qpn can be non-zero when there are other qp
that share the device. This patch fixes this issue by
using the correct qp num associated with the ep.

Signed-off-by: Shi Jin <sjina@amazon.com>