core: Introduce new inject calls that take buffer descriptor
Existing inject calls (for msg, tagged, rma, and atomic ops) don't provide
a way to supply the memory descriptor of the data buffer. This presents
difficulty when FI_HMEM is enabled. Providers either have to detect the
HMEM interface the buffer is allocated with, or declare the supported
inject_size be zero. Both may result in sub-optimal application performance.

Here a set of new inject calls are introduced. The difference of the new
calls vs the corresponding existing calls is the addition of a new
parameter "desc", like what have already existed in the send calls.
With the new calls, the provider is able to know the iface of the
data buffer and handle it properly.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>