prov/efa: Only do dmabuf reg when FI_MR_DMABUF is set
Application should use FI_MR_DMABUF API to pass the dmabuf
fd and offset to make Libfabric register the mr via
dmabuf. The only exception is for synapseai, beacuse
dmabuf is the only way to register Gaudi device buffer
and it was implemented before the FI_MR_DMABUF API.
Keep this behavior unchanged for compatibility.

Signed-off-by: Shi Jin <sjina@amazon.com>