prov/shm: Add unmap_region function
This function is mainly for the niche case where on progress_connreq
a peer is added to the map with its region needing to be mapped, and
then after mapping it, it's discovered that the newly mapped peer's
process died. In this case we need to unmap them and free any resources
that were opened for communicating with them.

With the creation of this function we can rework smr_map_del to use
it as common code. This requires changes to smr_av.c where smr_map_del
is called. smr_map_del is now an iterable function. This is to optimize
smr_map_cleanup to use ofi_rbmap_foreach to only cleanup peers that
exist.

Remove lock from map_to_region and unmap_region functions and require
lock acquirement before calling those functions. This is necessary because
on av removal path map will be double locked if the functions also process
locking the map. The map_to_region function is updated to mirror this
policy.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>