fabtests: fix complex fill cast
The previous implementation was trying to fill complex numbers with
multiple chars in the integ_alphabet to randomize the data. Unfortunately,
this resulted in many complex numbers becoming invalid numbers (NaN).
Because of that, the data validation check would fail as NaN != NaN
To fix this, when filling the data, set the real and imaginary parts as
a single char element so the individual part is valid and can be used
in the computation and check.
Because the implementation of complex types differs on unix and Windows
systems, this requires the addition of a new function that initializes
the individual components of the complex number instead of the number
as a whole.

This requires moving the long_double typedef into the osd.h headers to
help with the casting.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>