fabtests/pytest/efa: Loose assertion for read request counters
The current assertion check for send read wrs and bytes are
based on the exact calculation of benchmark data size + fabtests
control message size, which is too strict. Fabtests may change
control message sizes which should not impact this test.

This patch looses the assertion for read request counters which
make it have enough coverage for the benchmark data size.

Signed-off-by: Shi Jin <sjina@amazon.com>