contrib/intel/jenkins: Use new CI to build
New CI will be in charge of building everything that
needs to be built that it is in charge of running.
Libfabric, Fabtests, SHMEM.

New CI prefers a tarball to build from so we will checkout
scm from the plugin, tar it, and then feed it to new CI to use.

All builds will eventually move from middlewares to pre-build.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>