prov/verbs: Fix coverity issue about overflowed return value
The function `vrb_eq_xrc_recip_conn_event` was defined as size_t but
can return a nagative value on error which would overflow when casted
to int. Change the type to ssize_t.

Also change the return type of `vrb_eq_xrc_connected_event` to ssize_t
to be consistent with its caller and the function it calls.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>