[v1.22.x] prov/efa: Fix the read_bad_recv_status unit test
When device supports unsolicited write, efadv_wc_is_unsolicited is called in the error
path of IBV_WC_RECV as well, but we failed to setup its mock
in the unit test, it will cause a segfault inside the rdma-core
function when we don't really have a cqe, which is what the current unit test does.

This patch fixes this bug.

Signed-off-by: Shi Jin <sjina@amazon.com>
(cherry picked from commit 628dfa78866fe162695eff37624912d05fdbda9f)