prov/lnx: fix av strncpy
Fix compiler error with strncpy:

prov/lnx/src/lnx_av.c:272:4: warning: ‘strncpy’ specified bound 64 equals destination size [-Wstringop-truncation]
    strncpy(peer_prov->lpp_prov_name, prov_name, FI_NAME_MAX);

Both dest and srouce are FI_NAME_MAX but strncpy is complaining that it doesn't have space for the NULL
terminator if it needs to be truncated. The terminator should already be in the prov_name passed in.
Just turn this into a memcpy to make the compiler happy

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>