prov/efa: Decouple AV entry from endpoint
Before this change, the EFA AV entry contained a reference to
efa_rdm_peer which is specific to a given endpoint. This member also
prevented binding a single AV to multiple endpoints.

This change removes efa_rdm_peer from AV entry  by adding a hashmap
to the endpoint that maps fi_addr to efa_rdm_peer. And it also
enables multiple EFA endpoints to bind to the same AV.

Co-authored-by: Shi Jin <sjina@amazon.com>
Signed-off-by: Sai Sunku <sunkusa@amazon.com>