util/pingpong: close mr after ep close
pingpong doesn't support FI_MR_ENDPOINT today,
so the mr is associated with domain instead of ep.
It is unsafe to close mr before closing ep because
it can cause an EBUSY error when there are outstanding
recvs of the mr posted to the ep/qp. This patch fixes
this issue by moving the mr close after the ep close.

Signed-off-by: Shi Jin <sjina@amazon.com>