prov/util: fix FI_MULTI_RECV not set on FI_ECANCELED
When canceling a multi-recv entry, FI_MULTI_RECV was never
set as part of the error entry flags. Fix so that the user
can determine if the buffer is still in use.

Return -FI_EAGAIN on cancel if multi-recv buffer is in use.

Update fi_endpoint man page to describe fi_cancel() return value.

Signed-off-by: Jerome Soumagne <jerome.soumagne@hpe.com>