prov/shm: remove prefix from map inserts
EP names have the format fi_shm://<shm_name> to be consistent with
other provider formats. Since shared memory names cannot have slashes
in them, the shm provider skips this prefix when creating the shared
memory region. However, it does not do this when inserting the names
into the map which causes an issue when comparing names in the map
against region names.
This removes the prefix for map additions as well so that names match
in both data structures.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>