prov/efa: Rework the efa_cq unit tests
The current test_efa_cq* involves fi_send/fi_recv
operations which are not relevant to the cq operation
and make the test lose the value as unit tests.

This patch rework the efa_cq unit tests
to focus on  the cq read operation and adds
check for the returned cq entry.

Signed-off-by: Shi Jin <sjina@amazon.com>