prov/tcp: drop stale messages for new endpoint
When an communication instance is restarted, the new instance may still
receive 'stale' tagged replies that belong to the previous instance. In the
current implementation, it will disable the polling against the new endpoint
and then leave lots of normal messages unprocessed.

This PR is trying to fix the issue by dropping those messages with
unmatched tag and if the endpoint is entirely new.

DAOS ticket: https://daosio.atlassian.net/browse/DAOS-17103

Signed-off-by: Jinshan Xiong <jinshanx@google.com>