prov/efa: Clean up rxe map during rxe release
rxe can be inserted to rxe map when receiver
gets a multi-req pkts (medium, runting). When the
receive ep is closed early before the receive
completes, the stale entry in the rxe map will
cause an assertion error when destroying
the rxe_map buffer pool during ep close. This
patch fixes this issue by recording the rxe_map
insertion status in rxe, and removing the rxe
from the map during the rxe release.

Signed-off-by: Shi Jin <sjina@amazon.com>