prov/efa: Fix the max_msg_size reporting for efa-direct
Currently, efa-direct always return the max of rdma size
and mtu size when device rdma is supported. This is wrong
because when FI_RMA is not requested, only send/recv will
be used, and the max_msg_size should be returned as the
mtu size only. This patch fixes this issue by moving
the max_msg_size change to alter_direct_info based on
hints->caps.

Also adjust and introduces more unit tests.

Signed-off-by: Shi Jin <sjina@amazon.com>