include, prov: Change util_av lock to genlock
The util_av lock will be held in the fast path in the EFA provider. With the
correct threading and progress models
(FI_THREAD_DOMAIN/FI_THREAD_COMPLETION and FI_PROGRESS_CONTROL_UNIFIED),
this lock should also be a no-op. Similar to the ep_list_lock.

This commit changes the mutex to a genlock that becomes a no-op for the
correct threading and progress models and a mutex otherwise.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>