prov/efa: Fix the clean up issue for efa_util_prov
Commit b7b9dd6 makes efa_prov always returned
by EFA_INI, even when efa device is not available.
Then in `fi_fini` the efa_prov_finialize` will be
triggered on non-EFA platform. However, the current
efa_prov_finalize doesn't handle the case where
the resources were not created, and it will cause
a double free error.

This patch fixes this issue.

Signed-off-by: Shi Jin <sjina@amazon.com>