prov/efa: Second fix for efa-direct's max_msg_size
commit f3e26d61038ac8 introduces another bug
that the max_msg_size can be larger than
the the max_msg_size of efa_prov_info when
FI_RMA is requested: This will cause the
fi_endpoint() call failed when comparing
the user info and prov info.

This patch fixes this issue by making the
prov_info always use the largest max_msg_size,
and conditionally reduce the returned
user_info to the smaller max_msg_size
when FI_RMA is not requested.

Signed-off-by: Shi Jin <sjina@amazon.com>