man/fi_domain: Update connectionless EP disable
Previously, any resource management error would result in an endpoint
being disabled. Storage libfabric users have the expectation that
connectionless endpoints will not be disabled in an RDMA operation
results in an error. For example, if a client sends a server a bad
remote key and the server issues an RMA read/write to this key, by the
current resource management definition, the endpoint will be disabled.
RDMA failure to a single connectionless peer should not an impact
endpoints connectively to other peers.

Fix this by clarifying exactly which errors will cause a connectionless
endpoint to transition into a disabled state.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>