fabtests/efa: Add remote exit early test with post recv
When rdma read is available and message size >= 1M, long read or
runt read protocol is used and server is expected to get a cq entry
or cq error.
Otherwise, if long CTS is used and sender exits before sending CTS
data, receiver is expected to timeout after sending the CTS packet
without getting a cq entry or cq error.

Signed-off-by: Jessie Yang <jiaxiyan@amazon.com>