prov/efa: Minimize calls to efa_rdm_ep_get_peer in the CQ read path
This change is useful for two reasons
(1) The efa_rdm_ep_get_peer function accesses the AV and it's easier to
lock access to the AV when the call happens in fewer places
(2) With an implicit AV, the fi_addr will always be -1 for peers not
explicitly inserted to the AV. So efa_rdm_ep_get_peer will need to do
a reverse hashmap lookup with a significant performance cost for such
peers. So minimizing the number of calls to efa_rdm_ep_get_peer is more
performant.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>