prov/util: Only allocate MR key when needed
MR key is only needed for remote access. If remote access rights are not asked
for, MR registration should succeed even if key allocation fails. To prevent
key confliction with a future registration that needs the key, don't even try
to allocation the key if it is not needed.

A special key, FI_KEY_NOTAVAIL, is assigned for such case.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>