prov/efa: Use mutex locks for FI_THREAD_COMPLETE
The EFA provider does not support the FI_THREAD_COMPLETE locking
optimization. EFA attempts to lock the SRX lock in both the tx
and completion path. This behavior leads to an ofi_genlock assertion
being hit when a FI_THREAD_COMPLETE App runs with EFA compiled in
debug mode. Switch EFA to use locking in order to provide the correct
behavior for FI_THREAD_COMPLETE.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>