prov/tcp: fix flag and memory out of bound access
A change in ofi_verify_av_insert() is missed from the first PR.
Without that change, mercury won't be able to insert or change
flags for FI_FIREWALL_ADDR.

This PR also fixes an issue of out of bound memory access.
In xnet_ep_accept(), it allocates the size of data structure
`fi_eq_cm_entry{}` but `xnet_process_connected()` accesses it
as `xnet_cm_entry{}` that is larger in size. This is not introduced
by this series of PR but found it when I am doing the tests.

Signed-off-by: Jinshan Xiong <jinshanx@google.com>