prov/rxm: fix auto progress
When the application requests auto data progress, rxm needs to
start a thread to drive data progress. This is required not only
for atomic protocols but also for the rendezvous protocol.
The existing code was only enabling the automatic progress thread
when the application requested atomics.
This starts the correct thread function whenever auto progress
is requested, not just for atomics, allowing the rendezvous protocol
to pass in auto progress mode.
This also simplifies the checks to remove references to atomics
capabilities as well as the switch on the thread function since it
is now redundant.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>