prov/efa: Add lock around g_efa_domain_list
A race condition was discovered when initializing EFA using multiple
threads. The EFA Provider has a global variable called
g_efa_domain_list, which holds a list of all the EFA domains in a
fabric. The list is initialized in efa_prov.c in the fi_info path
(thread safe), and then used in domain open path (fork support, and
adding a domain to the list). The list needs to be protected by a
mutex so that we don't have a race condition of two threads trying
to modify the list at the same time. The libfabric man pages state
that all control interfaces will always be thread safe unless control
progress model is FI_PROGRESS_CONTROL_UNIFIED.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>